+++ src/main/kotlin/com/ravit/stash/shared/code/DependencyType.kt
@@ -1,0 +1,9 @@
+package com.ravit.stash.shared.code
+
+enum class DependencyType(
+    val critical: Boolean,
+) {
+    MONGODB(critical = true),
+    JIRA(critical = false),
+    SLACK(critical = false),
+}