+++ src/main/kotlin/com/ravit/stash/shared/code/HealthStatusType.kt
@@ -1,0 +1,7 @@
+package com.ravit.stash.shared.code
+
+enum class HealthStatusType {
+    UP,
+    DEGRADED,
+    DOWN,
+}